API Documentation

Operadic agent types

OperadicAgents.AbstractOperadicAgent — Type

Abstract supertype of all operadic agents. This is a dynamic structure which parametrizes dynamics of the agent, stores additional data required for the numerical simulation, and optionally logs its state at selected timesteps.

OperadicAgents.FreeAgent — Type

A container of operadic agents. Doesn't implement standalone evolutionary rule; delegates evolution to internal agents.

Implementing custom types

To implement a custom operadic agent type, you may want to use the convenience macro @oagent which supplies type fields expected (not required, though) by the interface.

Next step is to implement the required interface functions:

OperadicAgents._step! — Method

Project operadic agent's solution up to time t.

OperadicAgents._projected_to — Method

Return time to which operadic agent's evolution was projected.

OperadicAgents.getobservable — Method

getobservable(agent, args...)

Get operadic agent's observable.

Examples

getobservable(getagent(agent, "../model"), "observable_name")
getobservable(getagent(agent, "../model"), 1)

For a deeper integration of the agent type, you may specialize the following functions:

OperadicAgents._getparameters — Method

_getparameters(agent)

Retrieve parameter space of an operadic agent.

OperadicAgents._set_parameters! — Method

_set_parameters!

Mutate operadic agent's parameter space.

Examples

_set_parameters!(agent, Dict(:α=>1))
_set_parameters!(agent, [1., 2.])

OperadicAgents._draw — Method

Return plot of an operadic agent's state. Defaults to nothing.

OperadicAgents._reinit! — Method

Reinitialize the state of an operadic agent.

OperadicAgents._interact! — Method

Wake up an agent. See Opera.

OperadicAgents._prestep! — Method

Pre-step to a step call (e.g., projecting operadic agent's solution up to time t).

Common interface

Agent properties accessors

OperadicAgents.getname — Function

getname(agent)

Get operadic agent's name.

OperadicAgents.getuuid — Function

getuuid(agent)

Get operadic agent's uuid.

OperadicAgents.getparent — Function

getparent(agent)

Get operadic agent's parent.

OperadicAgents.inners — Function

inners(agent)

Get dictionary of operadic agent's inner agents. Follows name => agent format.

OperadicAgents.getopera — Function

Get operadic agent's Opera.

OperadicAgents.getdirectory — Function

Get operadic agent's directory. See also Opera.

OperadicAgents.getparameters — Function

getparameters(agent)

Retrieve operadic agents' (incl. inner agents, if applicable) parameter space.

OperadicAgents.set_parameters! — Function

set_parameters!(agent, parameters)

Assign operadic agent's parameters. Parameters are accepted in the form of a dictionary containing path => params pairs.

Examples

set_parameters!(agent, Dict("agent1/agent2" => Dict(:α=>1)))

Accessors

OperadicAgents.getobservable — Function

getobservable(agent, args...)

Get operadic agent's observable.

Examples

getobservable(getagent(agent, "../model"), "observable_name")
getobservable(getagent(agent, "../model"), 1)

OperadicAgents.gettimeobservable — Function

Get operadic agent's observable at a given time.

List observables observed by an agent and exported by an agent

OperadicAgents.in_observables — Function

Return a list of operadic agent's inner ports (subjective observables).

OperadicAgents.out_observables — Function

Return a list of operadic agent's outer ports (objective observables).

Solving & plotting

OperadicAgents.step! — Function

step!(agent, t=projected_to(agent))

Performs a single evolutionary step of the operadic hierarchy. To avoid frontrunning, solutions will be projected only up to time t. This is a two-phase step; the corresponding stepping functions are_prestep!andstep!`.

More particular behavior can be implemented using Opera protocol.

For custom operadic agents' types, it suffices to implement _step!.

Return values

Return true if all internal operadic agent's time horizon was reached. Else return the minimum time up to which the operadic agent's solution was projected.

OperadicAgents.simulate — Function

simulate(agent::AbstractOperadicAgent, max_t=Inf)::AbstractOperadicAgent

Solves an (initialized) operadic problem. Runs a loop until all the agents return true (reached simulation horizon) or nothing (delegated evolution), or until the simulation horizon reaches max_t. Avoids front-running.

Examples

sol = simulate(model)

OperadicAgents.draw — Function

Plot an operadic agent's state. See also _draw.

Paths

Implements path-like structure of operadic agents.

OperadicAgents.@glob_str — Macro

Returns a glob string to enable wildcard matching of operadic agents paths.

OperadicAgents.@uuid_str — Macro

Returns UUID object given a uuid string.

OperadicAgents.getagent — Function

Retrieve an operadic agent at path, relatively to agent.

getagent(a::AbstractOperadicAgent, uuid::UUID)

Get an operadic agent given its uuid.

Examples

getagent(a, UUID("2a634aad-0fbe-4a91-a605-bfbef4d57f95"))
getagent(a, uuid"2a634aad-0fbe-4a91-a605-bfbef4d57f95")

getagent(agent::AbstractOperadicAgent, path::AbstractString)

Get an operadic agent given its relative path.

Examples

getagent(a, "../agent")

getagent(agent::AbstractOperadicAgent, path::Union{Glob.FilenameMatch, Regex})

Get an operadic agent given a regex or glob string.

Examples

getagent(a, r"agent.*")
getagent(a, glob"**/agent/")

OperadicAgents.disentangle! — Function

disentangle!(agent)

Detach an agent from its parent. Optionally set remove_relpathrefs=false keyword to skip removing the relative pathrefs.

Examples

disentangle!(agent)

Opera, a dynamic structure to facilitate complex interactions

OperadicAgents.Opera — Type

Opera(uuid2agent_pairs...)

A dynamic structure that stores a priority queue of operadic interactions and contains a directory of operadic agents (dictionary of uuid => agent pairs).

Operadic Interactions

It is possible to schedule additional interactions within the complex; such actions are instances of AbstractOperaCall, and they are modeled as tuples (priority=0., call).

At the end of the topmost call to step!, the actions will be executed one-by-one in order of the respective priorities.

In particular, you may schedule interactions of two kinds:

@schedule agent priority=0, which will translate into a call _interact!(agent),

@schedule_call agent f(args...) priority=0 or @schedule_call agent x->ex priority=0,

which will translate into a call agent->f(agent, args...) or (x->ex)(agent), respectively.

See @schedule and @schedule_call.

they exist within a single step of the model and are executed after the calls to _prestep! and _step! finish.

See opera_enqueue!.

OperadicAgents.AbstractOperaCall — Type

Abstract opera interaction. See Opera.

OperadicAgents.AgentCall — Type

A scheduled call to an agent. See Opera.

OperadicAgents.opera_enqueue! — Function

Schedule an operadic interaction.

Operad

Defines general sums and products of operadic models.

OperadicAgents.:⊕ — Function

⊕(models::Vararg{AbstractOperadicAgent, N}; name)

Operadic sum of operadic models. Optionally specify resulting model's name.

By default, outputs an instance of FreeAgent.

Examples

⊕(m1, m2; name="diagram1") ⊕ ⊕(m3, m4; name="diagram2");

OperadicAgents.@sum — Macro

@sum models...

Perform an operadic sum of operadic models (flatten arguments to ⊕).

Examples

@sum m1 m2 m3 m4 # == ⊕(m1, m2, m3, m4)

Agent types macros

Supports convenient operadic agent subtyping.

OperadicAgents.@oagent — Macro

@oagent agent_name begin
    extra_fields...
end

Create a custom operadic agent type, and include fields expected by default interface methods (see FreeAgent).

Example

@oagent Molecule begin
    age::Float64
    birth_time::Float64
    kill_time::Float64

    mol::AbstractString
    profile::NTuple{N, Float64}

    sales::Float64
    df_sales::DataFrame
end

@oagent agent_name supertype begin
    extra_fields...
end

Create a custom operadic agent type, and include fields expected by default interface methods (see FreeAgent).

Example

@oagent SmallMolecule Molecule begin
    age::Float64
    birth_time::Float64
    kill_time::Float64

    mol::AbstractString
    profile::NTuple{N, Float64}

    sales::Float64
    df_sales::DataFrame
end

Walks

Walk operadic agents' hierarchy.

OperadicAgents.prewalk — Function

Applies f to each operadic agent, returning the result. Applies f to an agent before visiting its inners.

OperadicAgents.postwalk — Function

Applies f to each operadic agent, returning the result. Applies f to an agent before visiting its inners.

Utility functions

Add integration

OperadicAgents.add_integration — Function

add_integration(name)

Instantiate a package integration and add it to Julia's load path.

Available integrations

SciMLIntegration

AgentsIntegration

Expression wrappers

OperadicAgents.@wrap — Macro

@wrap(expr[, agent=:agent])

Turn expr into a function of agent, optionally specifying agent alias.

Examples

@wrap agent
@wrap ag ag

Observable accessor, interaction schedulers

OperadicAgents.@observables — Macro

@observables agent path:obs path:(obs1, obs2) path:[obs1, obs2]

Retrieve (a vector of) observables relative to agent.

Examples

@observables agent "../agent":"o"
@observables agent "../agent":("o1", "o2")

OperadicAgents.@schedule — Macro

@schedule agent priority=0

Schedule an interaction. Interactions are implemented within an instance Opera, sorted by their priorities. Internally, reduces to _interact!(agent).

See also Opera.

Examples

@schedule agent 1.

OperadicAgents.@schedule_call — Macro

@schedule agent call priority=0

Schedule an interaction (call). Interactions are implemented within an instance Opera, sorted by their priorities. Internally, the call=f(args...) expression will be transformed to an anonymous function agent -> f(agent, args...).

See also Opera.

Examples

@schedule agent f(t)

Flat representation

OperadicAgents.flatten — Function

flatten(root_agent)

Return flat representation of operadic agents hierarchy.