Agents.jl Integration in OperadicAgents.jl

The integration can be loaded as:

using OperadicAgents
add_integration(:AgentsIntegration); using AgentsIntegration

Operadic Wrap Types

AgentsIntegration.ABMAgent — Type

ABMAgent(name, abm; kwargs...)

Initialize ABMAgent, incl. hierarchy of ABM's agents.

Configure the evolutionary step, logging, and step size by keyword arguments below.

Arguments

agent_step!, model_step!: same meaning as in Agents.step!

in general, any kwarg accepted by Agents.run!, incl. adata, mdata

when, when_model: when to collect agents data, model data true by default, and performs data collection at every step if an AbstractVector, checks if t ∈ when; otherwise a function (model, t) -> ::Bool

step_size: how far the step advances, either a float or a function (model, t) -> size::Float64

tspan: solution horizon, defaults to (0., Inf)

AgentsIntegration.AAgent — Type

Operadic wrap for AbstractAgent type.

Operadic Bindings

AgentsIntegration.@get_oagent — Macro

@get_oagent model agent

Retrieve agent's operadic wrap from ABM's operadic wrap.

Examples

operadic_agent = @get_oagent abm_model abm_agent

AgentsIntegration.@o — Macro

@o operation

Operadic extension of add_agent!, kill_agent!.

Examples

@o add_agent!(model, 0.5)
@o disentangle!(agent, model)