SciML Integration in OperadicAgents.jl

The integration can be loaded as:

using OperadicAgents
add_integration(:SciMLIntegration); using SciMLIntegration

Model, Integrator, and Solution Wrap Types

SciMLIntegration.DiffEqAgent — Type

Operadic wrap of a SciML DE problem.

Operadic Bindings

SciMLIntegration.@get_oagent — Macro

@get_oagent p=:p key=:__oagent__

Retrieve operadic wrap, which is stored under key in parameter space p.

Examples

wrap = @get_oagent # assumes parameter space `p`, integrator's wrap ref `:__oagent__`
wrap = @get_oagent params
wrap = @get_oagent params :agent

Observables

SciMLIntegration.push_in_observables! — Function

push_in_observables!(a::DiffEqAgent, pairs...)

Register explicit in-ports of an operadic DiffEq model. Provide pairs path => observable, where observable may optionally be an iterable collection of observables' names.

Examples

push_out_observables!(deagent, path => observable, path => [observables...])

SciMLIntegration.push_out_observables! — Function

push_out_observables!(a::DiffEqAgent, pairs...)

Register explicit out-ports of an operadic DiffEq model. Enables aliasing of variable's positional index. That is, provide key => ix pair to alias ixth model's variable as key.

Examples

push_out_observables!(deagent, key => ix1, ix2)