API Documentation

Algebraic agent types

AlgebraicAgents.AbstractAlgebraicAgent — Type

Abstract supertype of all algebraic agents. This is a dynamic structure which parametrizes dynamics of the agent, stores additional data required for the numerical simulation, and optionally logs its state at selected timesteps.

AlgebraicAgents.FreeAgent — Type

A container of algebraic agents. Doesn't implement a standalone evolutionary rule; delegates evolution to internal agents.

AlgebraicAgents.FreeAgent — Method

FreeAgent(name, agents=[])

Initialize an algebraic agent. Optionally provide contained agents at the time of instantiation. See also entangle! and disentangle!.

Examples

FreeAgent("agent", [agent1, agent2])

Implementing custom types

To implement a custom algebraic agent type, you may want to use the convenience macro @aagent which supplies type fields expected (not required, though) by the interface.

Next step is to implement the required interface functions:

AlgebraicAgents._step! — Method

Project algebraic agent's solution up to time t.

AlgebraicAgents._projected_to — Method

Return time to which algebraic agent's evolution was projected.

AlgebraicAgents.getobservable — Method

getobservable(agent, args...)

Get algebraic agent's observable.

Examples

getobservable(getagent(agent, "../model"), "observable_name")
getobservable(getagent(agent, "../model"), 1)

For a deeper integration of the agent type, you may specialize the following functions:

AlgebraicAgents._getparameters — Method

_getparameters(agent)

Retrieve parameter space of an algebraic agent.

AlgebraicAgents._setparameters! — Method

_setparameters!

Mutate algebraic agent's parameter space.

Examples

_setparameters!(agent, Dict(:α=>1))
_setparameters!(agent, [1., 2.])

AlgebraicAgents._draw — Method

Return plot of an algebraic agent's state. Defaults to nothing.

AlgebraicAgents._reinit! — Method

Reinitialize the state of an algebraic agent.

AlgebraicAgents._interact! — Method

Wake up an agent. See Opera.

AlgebraicAgents._prestep! — Method

Pre-step to a step call (e.g., projecting algebraic agent's solution up to time t).

AlgebraicAgents._construct_agent — Method

Specialize the function to support convenience wrap initialization, dispatched on the arguments' types. See @wrap.

AlgebraicAgents._get_agent — Method

Extract algebraic wrap from obj.

Loading third-party package integrations

So far, integrations of DifferentialEquations.jl, Agents.jl, and AlgebraicDynamics.jl are provided.

Loading of the integrations is facilitated by Requires.jl; the integration will automatically be included once the respective third-party package is loaded.

For example,

using AlgebraicAgents
@isdefined DiffEqAgent

true

using AlgebraicAgents, DifferentialEquations
@wrap my_model ODEProblem((u, p, t) -> 1.01*u, [1/2], (0., 10.))

agent my_model with uuid ca192225 of type DiffEqAgent 
   custom properties:
   integrator: 
t: 0.0
u: 1-element Vector{Float64}:
 0.5

For plotting, you will want to load Plots as well. Nevertheless, function draw will inform you when necessary.

Common interface

Agent properties accessors

AlgebraicAgents.getname — Function

getname(agent)

Get algebraic agent's name.

AlgebraicAgents.getuuid — Function

getuuid(agent)

Get algebraic agent's uuid.

AlgebraicAgents.getparent — Function

getparent(agent)

Get algebraic agent's parent.

AlgebraicAgents.inners — Function

inners(agent)

Get dictionary of algebraic agent's inner agents. Follows name => agent format.

AlgebraicAgents.getopera — Function

Get algebraic agent's Opera.

AlgebraicAgents.getdirectory — Function

Get algebraic agent's directory. See also Opera.

AlgebraicAgents.getparameters — Function

getparameters(agent)

Retrieve algebraic agents' (incl. inner agents, if applicable) parameter space.

AlgebraicAgents.setparameters! — Function

setparameters!(agent, parameters)

Assign algebraic agent's parameters. Parameters are accepted in the form of a dictionary containing path => params pairs.

Examples

setparameters!(agent, Dict("agent1/agent2" => Dict(:α=>1)))

Accessors

AlgebraicAgents.getobservable — Function

getobservable(agent, args...)

Get algebraic agent's observable.

Examples

getobservable(getagent(agent, "../model"), "observable_name")
getobservable(getagent(agent, "../model"), 1)

AlgebraicAgents.gettimeobservable — Function

Get algebraic agent's observable at a given time.

List observables observed by an agent and exported by an agent

AlgebraicAgents.ports_in — Function

Return a list of algebraic agent's inner ports (subjective observables).

AlgebraicAgents.exposed_ports — Function

Return a list of algebraic agent's outer ports (objective observables).

Solving & plotting

AlgebraicAgents.step! — Function

step!(agent, t=projected_to(agent))

Performs a single evolutionary step of the algebraic hierarchy. To avoid frontrunning, solutions will be projected only up to time t. This is a two-phase step; the corresponding stepping functions are_prestep!andstep!`.

More particular behavior can be implemented using Opera protocol.

For custom algebraic agents' types, it suffices to implement _step!.

Return values

Return true if all internal algebraic agent's time horizon was reached. Else return the minimum time up to which the algebraic agent's solution was projected.

AlgebraicAgents.simulate — Function

simulate(agent::AbstractAlgebraicAgent, max_t=Inf)::AbstractAlgebraicAgent

Solves an (initialized) algebraic problem. Runs a loop until all the agents return true (reached simulation horizon) or nothing (delegated evolution), or until the simulation horizon reaches max_t. Avoids front-running.

Examples

sol = simulate(model)

AlgebraicAgents.draw — Function

Plot an algebraic agent's state. For internal implementation, see _draw.

Paths

Implements path-like structure of agents.

AlgebraicAgents.@glob_str — Macro

Returns a glob string to enable wildcard matching of algebraic agents paths.

AlgebraicAgents.@uuid_str — Macro

Returns UUID object given a uuid string.

AlgebraicAgents.getagent — Function

Retrieve an algebraic agent at path, relatively to agent.

getagent(a::AbstractAlgebraicAgent, uuid::UUID)

Get an algebraic agent given its uuid.

Examples

getagent(a, UUID("2a634aad-0fbe-4a91-a605-bfbef4d57f95"))
getagent(a, uuid"2a634aad-0fbe-4a91-a605-bfbef4d57f95")

getagent(agent::AbstractAlgebraicAgent, path::AbstractString)

Get an algebraic agent given its relative path.

Examples

getagent(a, "../agent")

getagent(agent::AbstractAlgebraicAgent, path::Union{Glob.FilenameMatch, Regex})

Get an algebraic agent given a regex or glob string.

Examples

getagent(a, r"agent.*")
getagent(a, glob"**/agent/")

Opera, a dynamic structure to facilitate complex interactions

AlgebraicAgents.Opera — Type

Opera(uuid2agent_pairs...)

A dynamic structure that stores a priority queue of algebraic interactions and contains a directory of algebraic agents (dictionary of uuid => agent pairs).

Algebraic Interactions

It is possible to schedule additional interactions within the complex; such actions are instances of AbstractOperaCall, and they are modeled as tuples (priority=0., call).

At the end of the topmost call to step!, the actions will be executed one-by-one in order of the respective priorities.

In particular, you may schedule interactions of two kinds:

@schedule agent priority=0, which will translate into a call _interact!(agent),

@schedule_call agent f(args...) priority=0 or @schedule_call agent x->ex priority=0,

which will translate into a call agent->f(agent, args...) or (x->ex)(agent), respectively.

See @schedule and @schedule_call.

they exist within a single step of the model and are executed after the calls to _prestep! and _step! finish.

See opera_enqueue!.

AlgebraicAgents.AbstractOperaCall — Type

Abstract opera interaction. See Opera.

AlgebraicAgents.AgentCall — Type

A scheduled call to an agent. See Opera.

AlgebraicAgents.opera_enqueue! — Function

Schedule an algebraic interaction.

Operations

Defines sums of agents.

AlgebraicAgents.:⊕ — Function

⊕(models::Vararg{AbstractAlgebraicAgent, N}; name)

Algebraic sum of algebraic models. Optionally specify resulting model's name.

By default, outputs an instance of FreeAgent.

Examples

⊕(m1, m2; name="diagram1") ⊕ ⊕(m3, m4; name="diagram2");

⊕(system1, system2; diagram=pattern, name)
⊕([system1, system2]; diagram=pattern, name)
⊕(Dict(:system1 => system1, :system2 => system2); diagram=pattern, name)

Apply oapply(diagram, systems...) and wrap the result as a GraphicalAgent.

⊕(system1, system2; diagram=pattern, name)
⊕([system1, system2]; diagram=pattern, name)
⊕(Dict(:system1 => system1, :system2 => system2); diagram=pattern, name)

Apply oapply(diagram, systems...) and wrap the result as a GraphicalAgent.

⊕(system1, system2; diagram=pattern, name)
⊕([system1, system2]; diagram=pattern, name)
⊕(Dict(:system1 => system1, :system2 => system2); diagram=pattern, name)

Apply oapply(diagram, systems...) and wrap the result as a GraphicalAgent.

⊕(system1, system2; diagram=pattern, name)
⊕([system1, system2]; diagram=pattern, name)
⊕(Dict(:system1 => system1, :system2 => system2); diagram=pattern, name)

Apply oapply(diagram, systems...) and wrap the result as a GraphicalAgent.

AlgebraicAgents.@sum — Macro

@sum models...

Perform an algebraic sum of algebraic models (flatten arguments to ⊕).

Examples

@sum m1 m2 m3 m4 # == ⊕(m1, m2, m3, m4)

Entangle and disentangle agents hierarchies.

AlgebraicAgents.entangle! — Function

entangle!(parent, agent)

Push an agent to the hierachy.

Examples

entagle!(parent, ancestor)

AlgebraicAgents.disentangle! — Function

disentangle!(agent)

Detach an agent from its parent. Optionally set remove_relpathrefs=false keyword to skip removing the relative pathrefs.

Examples

disentangle!(agent)

Agent types macros

Supports convenient agent subtyping.

AlgebraicAgents.@aagent — Macro

@aagent agent_name begin
    extra_fields...
end

Create a custom algebraic agent type, and include fields expected by default interface methods (see FreeAgent).

Example

@aagent Molecule begin
    age::Float64
    birth_time::Float64
    kill_time::Float64

    mol::AbstractString
    profile::NTuple{N, Float64}

    sales::Float64
    df_sales::DataFrame
end

@aagent agent_name supertype begin
    extra_fields...
end

Create a custom algebraic agent type, and include fields expected by default interface methods (see FreeAgent).

Example

@aagent SmallMolecule Molecule begin
    age::Float64
    birth_time::Float64
    kill_time::Float64

    mol::AbstractString
    profile::NTuple{N, Float64}

    sales::Float64
    df_sales::DataFrame
end

Walks

Walk agents' hierarchy.

AlgebraicAgents.prewalk — Function

Applies f to each algebraic agent, returning the result. Applies f to an agent before visiting its inners.

AlgebraicAgents.postwalk — Function

Applies f to each algebraic agent, returning the result. Applies f to an agent before visiting its inners.

Utility functions

Initialize wrap, extract wrap

AlgebraicAgents.@wrap — Macro

@wrap name args... kwargs...

A convenience macro to initialize an algebraic wrap, dispatched on the arguments' types. See _construct_agent.

Examples

@wrap "ode_agent" prob ODEProblem(f, u0, tspan)
@wrap ode_agent prob ODEProblem(f, u0, tspan)

AlgebraicAgents.@get_agent — Macro

@get_agent obj

Extract algebraic wrap from obj. Reduces to _get_agent, as overloaded by third-party packages integrations.

Examples

wrap = @get_agent params # for SciML integration
wrap = @get_agent model agent # for ABM integration

Observable accessor, interaction schedulers

AlgebraicAgents.@observables — Macro

@observables agent path:obs path:(obs1, obs2) path:[obs1, obs2]

Retrieve (a vector of) observables relative to agent.

Examples

@observables agent "../agent":"o"
@observables agent "../agent":("o1", "o2")

AlgebraicAgents.@schedule — Macro

@schedule agent priority=0

Schedule an interaction. Interactions are implemented within an instance Opera, sorted by their priorities. Internally, reduces to _interact!(agent).

See also Opera.

Examples

@schedule agent 1.

AlgebraicAgents.@schedule_call — Macro

@schedule agent call priority=0

Schedule an interaction (call). Interactions are implemented within an instance Opera, sorted by their priorities. Internally, the call=f(args...) expression will be transformed to an anonymous function agent -> f(agent, args...).

See also Opera.

Examples

@schedule agent f(t)

Flat representation

AlgebraicAgents.flatten — Function

flatten(root_agent)

Return flat representation of algebraic agents hierarchy.

Default plots for custom agent types

AlgebraicAgents.@draw_df — Macro

@draw_df T field

A macro to define _draw(T) such that it will plot a DataFrame stored under field.

Requires DataFrames and Plots to be available.

Examples

@draw_df my_type log # will plot `log` property (a DataFrame) of `my_type`'s instance