API Documentation

Algebraic agent types

AlgebraicAgents.AbstractAlgebraicAgent — Type

Abstract supertype of all algebraic agents. This is a dynamic structure which parametrizes dynamics of the agent, stores additional data required for the numerical simulation, and optionally logs its state at selected timesteps.

AlgebraicAgents.FreeAgent — Type

A container of algebraic agents. Doesn't implement standalone evolutionary rule; delegates evolution to internal agents.

Implementing custom types

To implement a custom algebraic agent type, you may want to use the convenience macro @oagent which supplies type fields expected (not required, though) by the interface.

Next step is to implement the required interface functions:

AlgebraicAgents._step! — Method

Project algebraic agent's solution up to time t.

AlgebraicAgents._projected_to — Method

Return time to which algebraic agent's evolution was projected.

AlgebraicAgents.getobservable — Method

getobservable(agent, args...)

Get algebraic agent's observable.

Examples

getobservable(getagent(agent, "../model"), "observable_name")
getobservable(getagent(agent, "../model"), 1)

For a deeper integration of the agent type, you may specialize the following functions:

AlgebraicAgents._getparameters — Method

_getparameters(agent)

Retrieve parameter space of an algebraic agent.

AlgebraicAgents._set_parameters! — Method

_set_parameters!

Mutate algebraic agent's parameter space.

Examples

_set_parameters!(agent, Dict(:α=>1))
_set_parameters!(agent, [1., 2.])

AlgebraicAgents._draw — Method

Return plot of an algebraic agent's state. Defaults to nothing.

AlgebraicAgents._reinit! — Method

Reinitialize the state of an algebraic agent.

AlgebraicAgents._interact! — Method

Wake up an agent. See Opera.

AlgebraicAgents._prestep! — Method

Pre-step to a step call (e.g., projecting algebraic agent's solution up to time t).

Common interface

Agent properties accessors

AlgebraicAgents.getname — Function

getname(agent)

Get algebraic agent's name.

AlgebraicAgents.getuuid — Function

getuuid(agent)

Get algebraic agent's uuid.

AlgebraicAgents.getparent — Function

getparent(agent)

Get algebraic agent's parent.

AlgebraicAgents.inners — Function

inners(agent)

Get dictionary of algebraic agent's inner agents. Follows name => agent format.

AlgebraicAgents.getopera — Function

Get algebraic agent's Opera.

AlgebraicAgents.getdirectory — Function

Get algebraic agent's directory. See also Opera.

AlgebraicAgents.getparameters — Function

getparameters(agent)

Retrieve algebraic agents' (incl. inner agents, if applicable) parameter space.

AlgebraicAgents.set_parameters! — Function

set_parameters!(agent, parameters)

Assign algebraic agent's parameters. Parameters are accepted in the form of a dictionary containing path => params pairs.

Examples

set_parameters!(agent, Dict("agent1/agent2" => Dict(:α=>1)))

Accessors

AlgebraicAgents.getobservable — Function

getobservable(agent, args...)

Get algebraic agent's observable.

Examples

getobservable(getagent(agent, "../model"), "observable_name")
getobservable(getagent(agent, "../model"), 1)

AlgebraicAgents.gettimeobservable — Function

Get algebraic agent's observable at a given time.

List observables observed by an agent and exported by an agent

AlgebraicAgents.in_observables — Function

Return a list of algebraic agent's inner ports (subjective observables).

AlgebraicAgents.out_observables — Function

Return a list of algebraic agent's outer ports (objective observables).

Solving & plotting

AlgebraicAgents.step! — Function

step!(agent, t=projected_to(agent))

Performs a single evolutionary step of the algebraic hierarchy. To avoid frontrunning, solutions will be projected only up to time t. This is a two-phase step; the corresponding stepping functions are_prestep!andstep!`.

More particular behavior can be implemented using Opera protocol.

For custom algebraic agents' types, it suffices to implement _step!.

Return values

Return true if all internal algebraic agent's time horizon was reached. Else return the minimum time up to which the algebraic agent's solution was projected.

AlgebraicAgents.simulate — Function

simulate(agent::AbstractAlgebraicAgent, max_t=Inf)::AbstractAlgebraicAgent

Solves an (initialized) algebraic problem. Runs a loop until all the agents return true (reached simulation horizon) or nothing (delegated evolution), or until the simulation horizon reaches max_t. Avoids front-running.

Examples

sol = simulate(model)

AlgebraicAgents.draw — Function

Plot an algebraic agent's state. See also _draw.

Paths

Implements path-like structure of algebraic agents.

AlgebraicAgents.@glob_str — Macro

Returns a glob string to enable wildcard matching of algebraic agents paths.

AlgebraicAgents.@uuid_str — Macro

Returns UUID object given a uuid string.

AlgebraicAgents.getagent — Function

Retrieve an algebraic agent at path, relatively to agent.

getagent(a::AbstractAlgebraicAgent, uuid::UUID)

Get an algebraic agent given its uuid.

Examples

getagent(a, UUID("2a634aad-0fbe-4a91-a605-bfbef4d57f95"))
getagent(a, uuid"2a634aad-0fbe-4a91-a605-bfbef4d57f95")

getagent(agent::AbstractAlgebraicAgent, path::AbstractString)

Get an algebraic agent given its relative path.

Examples

getagent(a, "../agent")

getagent(agent::AbstractAlgebraicAgent, path::Union{Glob.FilenameMatch, Regex})

Get an algebraic agent given a regex or glob string.

Examples

getagent(a, r"agent.*")
getagent(a, glob"**/agent/")

AlgebraicAgents.disentangle! — Function

disentangle!(agent)

Detach an agent from its parent. Optionally set remove_relpathrefs=false keyword to skip removing the relative pathrefs.

Examples

disentangle!(agent)

Opera, a dynamic structure to facilitate complex interactions

AlgebraicAgents.Opera — Type

Opera(uuid2agent_pairs...)

A dynamic structure that stores a priority queue of algebraic interactions and contains a directory of algebraic agents (dictionary of uuid => agent pairs).

Algebraic Interactions

It is possible to schedule additional interactions within the complex; such actions are instances of AbstractOperaCall, and they are modeled as tuples (priority=0., call).

At the end of the topmost call to step!, the actions will be executed one-by-one in order of the respective priorities.

In particular, you may schedule interactions of two kinds:

@schedule agent priority=0, which will translate into a call _interact!(agent),

@schedule_call agent f(args...) priority=0 or @schedule_call agent x->ex priority=0,

which will translate into a call agent->f(agent, args...) or (x->ex)(agent), respectively.

See @schedule and @schedule_call.

they exist within a single step of the model and are executed after the calls to _prestep! and _step! finish.

See opera_enqueue!.

AlgebraicAgents.AbstractOperaCall — Type

Abstract opera interaction. See Opera.

AlgebraicAgents.AgentCall — Type

A scheduled call to an agent. See Opera.

AlgebraicAgents.opera_enqueue! — Function

Schedule an algebraic interaction.

Operad

Defines general sums and products of algebraic models.

AlgebraicAgents.:⊕ — Function

⊕(models::Vararg{AbstractAlgebraicAgent, N}; name)

Algebraic sum of algebraic models. Optionally specify resulting model's name.

By default, outputs an instance of FreeAgent.

Examples

⊕(m1, m2; name="diagram1") ⊕ ⊕(m3, m4; name="diagram2");

AlgebraicAgents.@sum — Macro

@sum models...

Perform an algebraic sum of algebraic models (flatten arguments to ⊕).

Examples

@sum m1 m2 m3 m4 # == ⊕(m1, m2, m3, m4)

Agent types macros

Supports convenient algebraic agent subtyping.

AlgebraicAgents.@oagent — Macro

@oagent agent_name begin
    extra_fields...
end

Create a custom algebraic agent type, and include fields expected by default interface methods (see FreeAgent).

Example

@oagent Molecule begin
    age::Float64
    birth_time::Float64
    kill_time::Float64

    mol::AbstractString
    profile::NTuple{N, Float64}

    sales::Float64
    df_sales::DataFrame
end

@oagent agent_name supertype begin
    extra_fields...
end

Create a custom algebraic agent type, and include fields expected by default interface methods (see FreeAgent).

Example

@oagent SmallMolecule Molecule begin
    age::Float64
    birth_time::Float64
    kill_time::Float64

    mol::AbstractString
    profile::NTuple{N, Float64}

    sales::Float64
    df_sales::DataFrame
end

Walks

Walk algebraic agents' hierarchy.

AlgebraicAgents.prewalk — Function

Applies f to each algebraic agent, returning the result. Applies f to an agent before visiting its inners.

AlgebraicAgents.postwalk — Function

Applies f to each algebraic agent, returning the result. Applies f to an agent before visiting its inners.

Utility functions

Add integration

AlgebraicAgents.add_integration — Function

add_integration(name)

Instantiate a package integration and add it to Julia's load path.

Available integrations

SciMLIntegration

AgentsIntegration

Expression wrappers

AlgebraicAgents.@wrap — Macro

@wrap(expr[, agent=:agent])

Turn expr into a function of agent, optionally specifying agent alias.

Examples

@wrap agent
@wrap ag ag

Observable accessor, interaction schedulers

AlgebraicAgents.@observables — Macro

@observables agent path:obs path:(obs1, obs2) path:[obs1, obs2]

Retrieve (a vector of) observables relative to agent.

Examples

@observables agent "../agent":"o"
@observables agent "../agent":("o1", "o2")

AlgebraicAgents.@schedule — Macro

@schedule agent priority=0

Schedule an interaction. Interactions are implemented within an instance Opera, sorted by their priorities. Internally, reduces to _interact!(agent).

See also Opera.

Examples

@schedule agent 1.

AlgebraicAgents.@schedule_call — Macro

@schedule agent call priority=0

Schedule an interaction (call). Interactions are implemented within an instance Opera, sorted by their priorities. Internally, the call=f(args...) expression will be transformed to an anonymous function agent -> f(agent, args...).

See also Opera.

Examples

@schedule agent f(t)

Flat representation

AlgebraicAgents.flatten — Function

flatten(root_agent)

Return flat representation of algebraic agents hierarchy.