Agents.jl Integration

The integration can be loaded as:

Algebraic Wrap Types

AlgebraicAgents.ABMAgent — Type

ABMAgent(name, abm; kwargs...)

Initialize ABMAgent, incl. hierarchy of ABM's agents.

Configure the evolutionary step, logging, and step size by keyword arguments below.

Arguments

- `agent_step!`, `model_step!`: same meaning as in `Agents.step!`
- in general, any kwarg accepted by `Agents.run!`, incl. `adata`, `mdata`
- `when`, `when_model`: when to collect agents data, model data
true by default, and performs data collection at every step
if an `AbstractVector`, checks if `t ∈ when`; otherwise a function (model, t) -> ::Bool
- `step_size`: how far the step advances, either a float or a function (model, t) -> size::Float64
- `tspan`: solution horizon, defaults to `(0., Inf)`

AlgebraicAgents.AAgent — Type

Algebraic wrap for AbstractAgent type.

Algebraic Bindings

AlgebraicAgents.@a — Macro

@a operation

Algebraic extension of add_agent!, kill_agent!.

Examples

@a add_agent!(model, 0.5)
@a disentangle!(agent, model)