SciML Integration

Problem Wrap Type

AlgebraicAgents.DiffEqAgent — Type

DiffEqAgent(name, problem[, alg]; exposed_ports=nothing, ports_in=nothing, kwargs...)

Initialize DE problem algebraic wrap.

Keywords

exposed_ports: either nothing or a dictionary which maps keys to observable's positional index in u,

ports_in: either nothing or a vector of (subjective) observables,

other kwargs will be propagated to the integrator at initialization step.

Observables

AlgebraicAgents.push_ports_in! — Function

push_ports_in!(a::DiffEqAgent, pairs...)

Register explicit in-ports of an algebraic DiffEq model. Provide pairs path => observable, where observable may optionally be an iterable collection of observables' names.

Examples

push_ports_in!(deagent, path => observable, path => [observables...])

AlgebraicAgents.push_exposed_ports! — Function

push_exposed_ports!(a::DiffEqAgent, pairs...)

Register explicit out-ports of an algebraic DiffEq model. Enables aliasing of variable's positional index. That is, provide key => ix pair to alias ixth model's variable as key.

Examples

push_exposed_ports!(deagent, key => ix1, ix2)